Resolve S10P12A504-707 "Be api"

See merge request s10-webmobile1-sub2/S10P12A504!53

+++ backend/src/main/java/com/quiet/onterview/member/service/MemberService.java
@@ -1,6 +1,9 @@
 package com.quiet.onterview.member.service;
 
 import com.quiet.onterview.common.BaseException;
+import com.quiet.onterview.member.dto.request.MemberModifyNicknameRequest;
+import com.quiet.onterview.member.dto.request.MemberWithdrawRequest;
+import com.quiet.onterview.member.dto.response.MemberModifyNicknameResponse;
 import com.quiet.onterview.security.jwt.JwtTokenProvider;
 import com.quiet.onterview.member.dto.request.MemberLoginRequest;
 import com.quiet.onterview.member.dto.response.MemberLoginResponse;
@@ -53,11 +56,16 @@ public class MemberService {
         return memberMapper.memberToMemberLoginResponse(member, accessToken, refreshToken);
     }
 
-    public void modifyPassword(Long userId, MemberModifyPasswordRequest memberModifyPasswordRequest) {
+    public void modifyPassword(Long memberId, MemberModifyPasswordRequest memberModifyPasswordRequest) {
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new BaseException(ErrorCode.USER_NOT_EXISTS));
+        if(!passwordEncoder.matches(memberModifyPasswordRequest.getOriginal(), member.getPassword())) {
+            throw new BaseException(ErrorCode.PASSWORD_NOT_MATCHES);
+        }
         if(!isPasswordCorrespond(memberModifyPasswordRequest.getPassword(), memberModifyPasswordRequest.getConfirm())) {
             throw new BaseException(ErrorCode.PASSWORD_CANNOT_CONFIRM);
         }
-        updatePassword(userId, memberModifyPasswordRequest.getPassword());
+        updatePassword(memberId, memberModifyPasswordRequest.getPassword());
     }
 
     public MemberTokenResponse remakeMemberToken(String accessToken, String refreshToken) {
@@ -76,7 +84,12 @@ public class MemberService {
                 .build();
     }
 
-    public void withdrawUser(Long memberId) {
+    public void withdrawUser(Long memberId, MemberWithdrawRequest memberWithdrawRequest) {
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new BaseException(ErrorCode.USER_NOT_EXISTS));
+        if(!passwordEncoder.matches(memberWithdrawRequest.getPassword(), member.getPassword())) {
+            throw new BaseException(ErrorCode.PASSWORD_NOT_MATCHES);
+        }
         memberRepository.deleteById(memberId);
     }
 
@@ -94,5 +107,20 @@ public class MemberService {
 
     public int updatePassword(Long userId, String password) {
         return memberRepository.updatePassword(userId, passwordEncoder.encode(password));
+    }
+
+    public MemberModifyNicknameResponse modifyNickname(Long memberId,
+            MemberModifyNicknameRequest memberModifyNicknameRequest) {
+        String nickname = memberModifyNicknameRequest.getNickname();
+        if((nickname==null)) {
+            throw new BaseException(ErrorCode.REQUIRED_VALUE_NOT_EXISTS);
+        }
+        if(!isNicknameAvailable(nickname)) {
+            throw new BaseException(ErrorCode.NICKNAME_DUPLICATED);
+        }
+        memberRepository.updateNickname(memberId, nickname);
+        return MemberModifyNicknameResponse.builder()
+                .nickname(nickname)
+                .build();
     }
 }
